Created basic version of dynamically generating a calendar grid!
In this code snippet, we use Array.from() along with an arrow function to generate an array of elements based on the totalCells value. Let's break down how it works:

Array.from(): The Array.from() method creates a new array instance from an array-like or iterable object. In this case, we're using it to create an array of a specified length.

{ length: totalCells }: We pass an object as the first argument to Array.from(), where the length property is set to totalCells. This specifies the desired length of the generated array.

Arrow function: The second argument of Array.from() is an arrow function that determines how each element in the generated array should be rendered. It takes two parameters: the current element (not used in this case, denoted by _) and the index of the element.

(_, index) => ...: In the arrow function, we use _ as the parameter name to indicate that we're not using the current element within the function. We're only interested in the index value for generating the unique key prop.

JSX element: Inside the arrow function's body, we return the JSX element for each iteration. In this case, it's the div element with the grid-item class.

key={index}: We assign the key prop to each div element to ensure React can efficiently update and reconcile the elements when there are changes. In this case, we use the index value as the key since it's unique for each element.