Calander responsive to Backend
Edited Calander function to make it responsive to backend variables the user entered including days selected and hours inputed. (The only thing I couldn't figure out was the CSS styling. I was playing aroung with it, but couldn't find an appealing solution). Besides the styling the cell generation and hours generation works well.